requirements.txt
aiogram==3.15.0
anthropic==0.42.0
python-dotenv==1.0.1
python-docx==1.1.2
PyPDF2==3.0.1
openpyxl==3.1.5
openai==1.58.1
APScheduler==3.11.2
